requirements.txt
transformers==4.39.2
diffusers==0.23.1
accelerate==1.0.1
ml_collections==0.1.1
wandb>=0.16.0
safetensors>=0.4.1
peft==0.6.1
tensorboard
Pillow
tqdm
packaging
numpy
scipy
ftfy
regex
pytorch_fid
brisque